Training/LiveModelValidation/requirements.txt
tensorflow>=2.19.0
opencv-python>=4.5.0
numpy>=1.21.0
python-dotenv>=0.21.0 